store dto들 분리
