+++ domain-module/src/main/java/dto/store/seller/request/PatchDto.java
@@ -1,8 +1,12 @@
 package dto.store.seller.request;
 
+import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+@NoArgsConstructor
+@AllArgsConstructor
 @Getter
 @Setter
 public class PatchDto {
