+++ domain-module/src/main/java/dto/store/seller/request/PostDto.java
@@ -1,0 +1,16 @@
+package dto.store.seller.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+@Setter
+public class PostDto {
+
+    private String saleMatters; // 가게 공지사항
+
+}
